+++ src/main/java/fr/iutinfo/studiesWar/models/Controle.java
@@ -15,9 +15,9 @@ public class Controle {
 	private Map<Personnage, Note> notes=new HashMap<Personnage, Note>();
 	private ArrayList<Evenement> events=new ArrayList<Evenement>();
 	private Partie partie;
-	private byte date;
+	private int date;
 	
-	public Controle(String s,Partie p,byte date){
+	public Controle(String s, Partie p, int date){
 		this.matiere=s;
 		this.partie=p;
 		this.date=date;
@@ -91,7 +91,7 @@ public class Controle {
 		return matiere;
 	}
 
-	public byte getDate() {
+	public int getDate() {
 		return date;
 	}
