+++ src/main/java/fr/iutinfo/studiesWar/models/effet/EchangeCopie.java
@@ -6,8 +6,8 @@ import fr.iutinfo.studiesWar.models.Personnage;
 public class EchangeCopie extends Effet{
 	private Personnage personnage;
 
-	public EchangeCopie(byte dD, byte dF,Personnage p) {
-		super(dD, dF, (byte) 4);
+	public EchangeCopie(int dD, int dF,Personnage p) {
+		super(dD, dF, 4);
 		this.personnage = p;
 	}
